# modelcraft-backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ModelCraft REST API
  description: |
    REST API for tenant management (Auth & Organization) in the ModelCraft platform.

    Business domain APIs (Projects, Models, Clusters, Enums) are served exclusively
    via GraphQL at /org/{orgName}/design/graphql. This OpenAPI spec covers only identity and
    tenant-management endpoints.
  version: 3.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
paths:
  /api/auth/register:
    post:
      operationId: register
      summary: Register user and initialize profile atomically
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful (user + profile created)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid input (phone format, password too short)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthInvalidInputError'
        '409':
          description: phone or userName already exists
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PhoneAlreadyRegisteredError'
                  - $ref: '#/components/schemas/UserNameAlreadyRegisteredError'
        '500':
          description: Profile init failed or internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
  /api/auth/login:
    post:
      operationId: login
      summary: Login with phone number and password
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid input (phone format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthInvalidInputError'
        '401':
          description: Authentication failed (phone not found or wrong password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
...
  /api/auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh tokens using httpOnly cookie (rotation)
      tags:
        - Auth
      security: []
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid, expired, or reused refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BaseResponse:
      type: object
      required:
        - requestId
      properties:
        requestId:
          type: string
          description: Unique request trace ID
          example: req_550e8400e29b41d4
    SystemError:
      type: object
      required:
        - requestId
        - error
      properties:
        requestId:
          type: string
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - SYSTEM_ERROR
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    AuthInvalidInputError:
      type: object
      required:
        - requestId
        - error
      properties:
        requestId:
          type: string
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - PARAM_INVALID.AUTH
            message:
              type: string
    PhoneAlreadyRegisteredError:
      type: object
      required:
        - requestId
        - error
      properties:
        requestId:
          type: string
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - CONFLICT.USER
            message:
              type: string
    AuthenticationError:
      type: object
      required:
        - requestId
        - error
      properties:
        requestId:
          type: string
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - AUTHENTICATION_FAILED
            message:
              type: string
    RegisterRequest:
      type: object
      required:
        - phone
        - password
        - userName
      properties:
        phone:
          type: string
          description: 11-digit mainland China phone number
          pattern: ^1[3-9]\d{9}$
          example: '13800138000'
        password:
          type: string
          description: Password (minimum 8 characters)
          minLength: 8
          example: mypassword123
        userName:
          type: string
          description: User name (3-32 chars, start with letter/underscore/hyphen, then alphanumeric/underscore/hyphen)
          pattern: ^[a-zA-Z_-][a-zA-Z0-9_-]{2,31}$
          minLength: 3
          maxLength: 32
          example: john_doe
    RegisterResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - userId
            - orgName
            - profile
          properties:
            userId:
              type: string
              description: Created user UUID
            orgName:
              type: string
              description: Created personal organization slug
            profile:
              $ref: '#/components/schemas/RegisterProfileSnapshot'
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Login identifier - can be phone number or username depending on identifierType
          example: '13800138000'
        identifierType:
          type: string
          description: 'Type of identifier: PHONE (default) or USERNAME'
          enum:
            - PHONE
            - USERNAME
          default: PHONE
        password:
          type: string
          description: User password
          example: mypassword123
        phone:
          type: string
          description: '[Deprecated] 11-digit mainland China phone number. Use identifier + identifierType instead.'
          pattern: ^1[3-9]\d{9}$
          example: '13800138000'
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - userId
            - accessToken
            - expiresIn
          properties:
            userId:
              type: string
              description: ModelCraft internal user UUID
            userName:
              type: string
              description: User display name
            orgName:
              type: string
              description: Primary organization name (first org the user belongs to, if any)
            accessToken:
              type: string
              description: Short-lived JWT access token
            expiresIn:
              type: integer
              description: Access token TTL in seconds
    RefreshResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - accessToken
            - expiresIn
          properties:
            accessToken:
              type: string
              description: New short-lived JWT access token
            expiresIn:
              type: integer
              description: Access token TTL in seconds
...
    RegisterProfileSnapshot:
      type: object
      required:
        - id
        - userId
        - nickname
      properties:
        id:
          type: string
        userId:
          type: string
        nickname:
          type: string
          description: default generated nickname, e.g. user_A1B2C3
        avatarUrl:
          type: string
        bio:
          type: string
    UserNameAlreadyRegisteredError:
      type: object
      required:
        - requestId
        - error
      properties:
        requestId:
          type: string
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - CONFLICT.USER
            message:
              type: string